Abstract doc builder
